podido dto comienzo correcto

+++ backend/src/main/java/com/servipet/backend/Pedido/Repositorio/RepositorioEstadoEntrega.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RepositorioEstadoEntrega extends JpaRepository<EstadoEntrega, Byte> {
+public interface RepositorioEstadoEntrega extends JpaRepository<EstadoEntrega, Integer> {
 }
